# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JamberTech E-SERVICES SMM Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/balance:
    patch:
      operationId: addBalance
      tags: [users]
      summary: Add balance to user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBalanceBody"
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (Admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all active services
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
...
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        balance:
          type: number
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - balance
        - role
        - createdAt
...
    AddBalanceBody:
      type: object
      properties:
        userId:
          type: integer
        amount:
          type: number
      required:
        - userId
        - amount
...
    Service:
      type: object
      properties:
        id:
          type: integer
        providerServiceId:
          type: string
        providerId:
          type: integer
        providerName:
          type: ["string", "null"]
        name:
          type: string
        category:
          type: string
        type:
          type: string
        originalRate:
          type: number
        sellRate:
          type: number
        min:
          type: integer
        max:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - providerServiceId
        - providerId
        - name
        - category
        - type
        - originalRate
        - sellRate
        - min
        - max
        - status
        - createdAt
    UpdateServiceBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        sellRate:
          type: number
        status:
          type: string
          enum: [active, inactive]
...