# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech E-SERVICES SMM Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceStatus } from "./serviceStatus";

export interface Service {
  id: number;
  providerServiceId: string;
  providerId: number;
  /** @nullable */
  providerName?: string | null;
  name: string;
  category: string;
  type: string;
  originalRate: number;
  sellRate: number;
  min: number;
  max: number;
  status: ServiceStatus;
  createdAt: Date;
}