# lib/api-zod/src/generated/types/updateServiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech E-SERVICES SMM Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateServiceBodyStatus } from "./updateServiceBodyStatus";

export interface UpdateServiceBody {
  name?: string;
  category?: string;
  sellRate?: number;
  status?: UpdateServiceBodyStatus;
}