# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JamberTech E-SERVICES SMM Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  balance: number;
  role: UserRole;
  createdAt: Date;
}